+++ submission/RELEASES.md
@@ -171,3 +171,11 @@ Based on:
 - Speakeasy CLI 1.443.0 (2.461.4) https://github.com/speakeasy-api/speakeasy
 ### Generated
 - [python v1.9.3] submission
+
+## 2024-11-26 00:27:48
+### Changes
+Based on:
+- OpenAPI Doc 1.0.0 https://docs.api.epilot.io/submission-api.yaml
+- Speakeasy CLI 1.446.0 (2.462.0) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [python v1.10.0] submission